Fix compilation errors: mesh_event_t type, format string, and RSSI type mismatch

+++ uc/buttonsMeshIDF/src/mesh_comm.c
@@ -188,7 +188,7 @@ void status_report_task(void *arg)
                 uint32_t free_heap = esp_get_free_heap_size();
                 
                 // Get RSSI from parent connection
-                int8_t rssi = 0;
+                int rssi = 0;
                 esp_wifi_sta_get_rssi(&rssi);
                 
                 // Check for low heap
